requirements.txt
torch==1.3
torchvision==0.4
numpy
pillow
scipy
matplotlib
